requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
CacheControl==0.14.3
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
dash==3.1.1
dash-bootstrap-components==2.0.3
dash-chat==0.3.0
dash-core-components==2.0.0
dash-extensions==2.0.4
dash-html-components==2.0.0
dash-iconify==0.1.2
dash_ag_grid==31.3.1
dash_mantine_components==2.1.0
dataclass-wizard==0.35.0
db-dtypes==1.4.3
EditorConfig==0.17.1
firebase_admin==7.0.0
Flask==3.1.1
Flask-Caching==2.3.1
google-api-core==2.25.1
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.35.1
google-cloud-bigquery-storage==2.32.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-genai==1.27.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsbeautifier==1.15.4
MarkupSafe==3.0.2
more-itertools==10.7.0
msgpack==1.1.0
narwhals==2.0.1
nest-asyncio==1.6.0
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pipdeptree==2.28.0
plotly==6.2.0
proto-plus==1.26.1
protobuf==6.31.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
retrying==1.3.4
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
tenacity==8.5.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
websockets==15.0.1
Werkzeug==3.1.3
zipp==3.23.0
